dev_requirements.txt
-r requirements.txt
django-extensions==1.5.5
fake-factory==0.5.2
Werkzeug
